clean: refactoring SignUp and SignIn
